Separate information view into Earthquake.xaml and Userquake.xaml
