+++ WpfClient/Pages/Informations/Earthquake.xaml
@@ -1,0 +1,87 @@
+﻿<Page
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:ui="http://schemas.modernwpf.com/2019"
+      xmlns:local="clr-namespace:WpfClient.Pages.Informations"
+      xmlns:EPSPDataView="clr-namespace:WpfClient.EPSPDataView" x:Class="WpfClient.Pages.Informations.Earthquake"
+      Background="{DynamicResource SystemControlPageBackgroundAltHighBrush}"
+      Foreground="{DynamicResource SystemControlPageTextBaseHighBrush}"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="712"
+      Title="Earthquake">
+    <Page.Resources>
+        <local:DetailDescriptionTemplateSelector x:Key="DetailTemplateSelector" />
+
+        <DataTemplate x:Key="DetailHeaderItem">
+            <TextBlock Text="{Binding Text}" Style="{StaticResource SubtitleTextBlockStyle}" />
+        </DataTemplate>
+
+        <DataTemplate x:Key="DetailPrefectureItem">
+            <TextBlock Text="{Binding Text}" Margin="0,16,0,0" Style="{StaticResource BaseTextBlockStyle}" />
+        </DataTemplate>
+
+        <DataTemplate x:Key="DetailScaleItem">
+            <TextBlock Text="{Binding Text}" Style="{StaticResource BaseTextBlockStyle}" />
+        </DataTemplate>
+
+        <DataTemplate x:Key="DetailNameItem">
+            <TextBlock Text="{Binding Text}" Margin="0,0,0,4" Style="{StaticResource BodyTextBlockStyle}" />
+        </DataTemplate>
+    </Page.Resources>
+    <Page.DataContext>
+        <EPSPDataView:EPSPQuakeView/>
+    </Page.DataContext>
+
+    <Grid>
+        <Image Source="{Binding BitmapImage}" Stretch="Uniform" />
+        <Canvas>
+            <Image Source="/Resources/MapOverlays/quake_note.png" Width="40" Canvas.Bottom="8" Canvas.Right="8"/>
+            <Image Source="/Resources/MapOverlays/description.png" Height="8" Canvas.Bottom="0" Canvas.Left="0" />
+        </Canvas>
+        <Canvas VerticalAlignment="Top" Height="40" Background="{DynamicResource SystemControlPageBackgroundAltMediumBrush}">
+            <ui:SimpleStackPanel Orientation="Horizontal" Spacing="16" Margin="4">
+                <ui:SimpleStackPanel Orientation="Vertical">
+                    <TextBlock Text="日時" Style="{StaticResource CaptionTextBlockStyle}" />
+                    <Label Content="{Binding EventArgs.OccuredTime}" FontWeight="Bold" />
+                </ui:SimpleStackPanel>
+                <ui:SimpleStackPanel Orientation="Vertical">
+                    <TextBlock Text="震源" Style="{StaticResource CaptionTextBlockStyle}" />
+                    <Label Content="{Binding EventArgs.Destination}" FontWeight="Bold" />
+                </ui:SimpleStackPanel>
+                <ui:SimpleStackPanel Orientation="Vertical">
+                    <TextBlock Text="深さ" Style="{StaticResource CaptionTextBlockStyle}" />
+                    <Label Content="{Binding EventArgs.Depth}" FontWeight="Bold" />
+                </ui:SimpleStackPanel>
+                <ui:SimpleStackPanel Orientation="Vertical">
+                    <TextBlock Text="M" Style="{StaticResource CaptionTextBlockStyle}" />
+                    <Label Content="{Binding EventArgs.Magnitude}" FontWeight="Bold" />
+                </ui:SimpleStackPanel>
+                <ui:SimpleStackPanel Orientation="Vertical">
+                    <TextBlock Text="津波" Style="{StaticResource CaptionTextBlockStyle}" />
+                    <Label Content="{Binding TsunamiText}"
+                                       FontWeight="Bold"
+                                       Foreground="{Binding TsunamiForeground}"
+                                       />
+                </ui:SimpleStackPanel>
+            </ui:SimpleStackPanel>
+        </Canvas>
+        <Expander Visibility="Visible" ExpandDirection="Left"
+                                  HorizontalAlignment="Right"
+                                  Background="{DynamicResource SystemControlPageBackgroundAltMediumBrush}"
+                                  IsExpanded="False" Header="詳細" MaxWidth="320" Margin="0,40,0,0">
+            <ScrollViewer Width="240" Padding="4">
+                <ItemsControl
+                                ItemsSource="{Binding DetailItemViewList}"
+                                ItemTemplateSelector="{StaticResource DetailTemplateSelector}">
+                    <ItemsControl.ItemsPanel>
+                        <ItemsPanelTemplate>
+                            <ui:SimpleStackPanel Spacing="4" />
+                        </ItemsPanelTemplate>
+                    </ItemsControl.ItemsPanel>
+                </ItemsControl>
+            </ScrollViewer>
+        </Expander>
+    </Grid>
+</Page>
